moved some fuctions to utils.py
